Notes and pretty card display
Filled out the readme.md for builder with some ideas for the future. Scratch is a stab at printing out cards  in the console with pretty markup.

It's getting involved, so I'm thinking it will migrate to a 'logger' class of sorts that can handle the details about displaying stuff in the console. Gonna snapshot commit and spin this off.

(The loop eventually breaks due to selecting a card twice since popping modifies the referenced object)