Argument 'measure'
This spits out a report that describes, field by field, the longest values and how long they are. This report will be helpful for constructing the database and the display_card output

I'm now constructing the formatter with a pair of dictionaries that lets you call the formatted field by field name alone. Pretty slick. (Also I did this commit on the light rail. Sweet)