Names normalization treatment
New names table that bridges off of cards. These two tables working together can construct the multi-value names field for API responses.

Could use this as a bridge table relying on another table that actually holds values, but there is no reason to do so with this field. Might do it on color.