Database_Service Cleanup
It's soooo pretty!

Created a 'query' method that holds all the try-else statements that were all over this service. Made a custom Writer error for it so  I can see the query that caused problems. Moved formatting of the query into the actual query methods which makes a lot of sense to me. I really like how this is looking (add_card still has some outdated stuff that ought to be changed, and I'm not sure how I feel about card_formatter.py at this point in time)