All remaining fields
Destroyed card_formatter and replaced with translator_service that only works for the DB. Discarded the measure and random portions of the script as they are no longer useful.

All fields are now added, but I omitted the vanguard fields because who cares and there may be additional stuff in the extras that I'll want down the line. In any case, it's time to move on to non-db-builder stuff.