Plugged in SpringBoot
Very basic port of the simple API I built in SpringBoot before. The pages were tinkered with a little bit, and a Card object + DAO was stubbed to get to the point where I can see an API response of id and name. That is done, so now it's time to turn it into it's own entity!

I could have sworn I created this new with Gradle rather than Maven, but somewhere along the line I noticed I'm actually on Maven and don't feel like backing it out. 

No commitment as of yet to the front-end solution, but I'm going to use jQuery for this phase of using the front end to peak at output. Don't spend time on 'finalizing' front-end shenanigans just yet.