Bootstrap nav-links
Adding in nav-link tabs for perusing API calls with multiple results. Clicking on a tab 'selects' it, but will also need to switch the display pane (which has to be moved) for the card that is selected.

The tabs right now appear on pageload, but whether or not to show them and how many appear needs to be baked into the javascript.