It's doing the thing, needs cleanup
The tabination of API results on the API browser is working. I've toyed with some organizational stuff, so the whole thing is out of whack and needs some attention on the class naming approach and such. The separation of js files is a little weird and will need to be ironed out as well.