Multi-Response-Display in the Flesh
Cleaned up the jQuery shenanigans so they are a bit easier to work with. Not splitting hairs over duplicated selectors at this time, but the result is a lot more organized.

I guess this is a checkpoint commit since there's a few more things I want to do before pulling this into Master.