This won't work
Replacing ?s with user-input is only for variables in the 'WHERE' portion of the SQL statement, which means I can't do what I want to do with it easily. Gonna drop this off and come up with a better way to do it.

(There was a botched merge around these parts which this commit may have picked up.)