Starting down the path of enabling search
As of right now, the top search field for apis.html will fire a cards?name=xxx API call which will eventually feed into active results being listed beneath the search field. I'm worried about a race condition and how this will play out when complete, but for now it's stubbed and logs to the console.