Fix for Variations construction within Database Builder
Variations coming from MTGJson only include additional printings in the same set. I want this to include all other printings of the card with the same name.

This is constructed after the overall database is built in order to benefit from having complete information at that point in time. It is feasible to accomplish this within the database build script, so revisit doing so down the line.