Moved Variations into main loop
I had originally constructed the Variations builder to be a separate loop after the initial was completed. That's not necessary, and now it all happens in the first loop.

Alongside this, there's some cleanup going on in here where I moved the building of the normalized tables into a single function that gets called with different params. Names table got an update from card -> card_id in order to conform with other tables like it.