Primary loop moved to main.py
Rather than let databaseService hog all the glory, I've moved the main loop to the primary script that is run. This let me do some swanky things with arguments and conditionals and the like. DatabaseService for example delegates the decision on whether or not to print verbose output to the query method thanks to the constructor recording the invocation at the start of the script.