Improved Output
Prepping for the inclusion of Images in the main loop. I'm playing with utilizing a more generic 'alternate between action and highlight' but this doesn't allow for error output.

Overall, the output for the loop will probably need to be revisited. It's all nice and pretty and stuff, but without the option to interrupt to share information during runtime it's not doing all it can do.