Set endpoint supports random selection
When the ID is 0, the endpoint will find the count of sets, generate an int between 0 and the count and then display that set.

For ORM, this is done on the Java side since ORM doesn't really like ordering and limiting as was used with the straight up JDBC approach.