Multiple Data Sources OOP Attempt
It occurred to me while building the scryfall service, that this could be extended from a parent datasource. Eventually, it also occurred to me that the database could be treated in the same way. Naturally, there would need to be a separate object that works with writing to the database, and I'm curious if that will be able to be done with how the connections and cursors work. For now, this is pretty cool and I want to see where it goes.