Endpoint returns basic Card
The trick was instantiating the list as an empty array list so that the .add() method was properly appending. As of right now, I'm seeing the card data reported through postman.

Next up is handling errors correctly such that the endpoint reports the error (with the intent that the front-end will eventually play nicely with these errors)

Also, because we're delivering a dynamic payload, I need to put some effort into figuring out how to handle for that. As of now, the endpoint expects to be delivering a list of Cards, which won't always be the case.